requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
db-dtypes==1.2.0
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-bigquery==3.17.1
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
pandas-gbq==0.20.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydata-google-auth==1.8.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
streamlit-aggrid==0.3.4.post3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
